Another batch moved
