+++ 2dlib/ddgr.h
