+++ 2dlib/grdefs.h
