+++ bitmap/CMakeLists.txt
@@ -30,4 +30,7 @@ target_include_directories(bitmap PUBLIC
   $<BUILD_INTERFACE:
     ${PROJECT_SOURCE_DIR}/ddio
   >
+  $<BUILD_INTERFACE:
+    ${PROJECT_SOURCE_DIR}/2dlib
+  >
 )
