+++ dd_video/CMakeLists.txt
@@ -31,4 +31,7 @@ target_include_directories(dd_video PUBLIC
   $<BUILD_INTERFACE:
     ${PROJECT_SOURCE_DIR}/linux
   >
+  $<BUILD_INTERFACE:
+    ${PROJECT_SOURCE_DIR}/2dlib
+  >
 )
