+++ libacm/adecode.h
