+++ libacm/aencode.h
