+++ networking/CMakeLists.txt
@@ -32,4 +32,7 @@ target_include_directories(networking PUBLIC
   $<BUILD_INTERFACE:
     ${PROJECT_SOURCE_DIR}/ddio
   >
+  $<BUILD_INTERFACE:
+    ${PROJECT_SOURCE_DIR}/2dlib
+  >
 )
