+++ physics/CMakeLists.txt
@@ -28,4 +28,7 @@ target_include_directories(physics PUBLIC
   $<BUILD_INTERFACE:
     ${PROJECT_SOURCE_DIR}/ddio
   >
+  $<BUILD_INTERFACE:
+    ${PROJECT_SOURCE_DIR}/2dlib
+  >
 )
