+++ ui/CMakeLists.txt
@@ -43,5 +43,8 @@ target_include_directories(ui PUBLIC
   $<BUILD_INTERFACE:
     ${PROJECT_SOURCE_DIR}/dd_video
   >
+  $<BUILD_INTERFACE:
+    ${PROJECT_SOURCE_DIR}/2dlib
+  >
 )
